Require admin privs for $rm

+++ cmd/rm.fish
@@ -1,7 +1,9 @@
 . lib/file.fish
 set file (file_file $rest)
 if test -n $file -a -f $file
-    rm $file
+    if contains $nick $ADMINS
+        rm $file
+    end
 else
     msg $chan $nick': no such file'
 end
